requirements.txt
certifi==2021.5.30
chardet==4.0.0
idna==2.10
Pillow==8.3.1
PyYAML==5.4.1
requests==2.25.1
selenium==3.141.0
urllib3==1.26.6
